cleaning up changes that will be moved to diff branches
