--- .github/deploy.yml
@@ -1,25 +1,0 @@
-# name: Upload Python Package to pipy
-
-# on:
-#   release:
-#     types: [published]
-
-# jobs:
-#   deploy:
-#     runs-on: ubuntu-latest
-#     steps:
-#       - uses: actions/checkout@v4
-#       - name: Set up Python
-#         uses: actions/setup-python@v4
-#         with:
-#           python-version: '3.x'
-#       - name: Install dependencies
-#         run: |
-#           python -m pip install --upgrade pip
-#           pip install build
-#       - name: Build package
-#         run: python -m build
-      # - name: Publish package
-      #   uses: pypa/gh-action-pypi-publish@v1
-      #   with:
-      #     password: ${{ secrets.PYPI_API_TOKEN }}
