--- .github/utils/test.yml
@@ -1,1 +1,0 @@
-print("lgtm")
