+++ scripts/test_efficiency.py
@@ -54,49 +54,6 @@ test_input_dir = test_input_dir
 #     flexible_collection.find_flow_components_object()
 #     flexible_collection.calculate_flows_obj()
 
-
-# #@profile
-# def test_small_tree_proj(small_tree, ez_projection):
-#     ez_projection.numba_project_cylinders("XY")
-#     small_tree.project_cylinders("XY")
-#     assert 1 == 1
-
-
-# #@profile
-# def test_numba_small_tree_proj(small_tree, ez_projection):
-#     ez_projection.numba_project_cylinders("XY")
-#     small_tree.numba_project_cylinders("XY")
-#     assert 1 == 1
-
-
-# #@profile
-# def test_happy_proj(happy_path_projection, ez_projection):
-#     ez_projection.numba_project_cylinders("XY")
-#     happy_path_projection.project_cylinders("XZ")
-#     assert 1 == 1
-
-
-# #@profile
-# def test_numba_happy_proj(happy_path_projection, ez_projection):
-#     ez_projection.numba_project_cylinders("XY")
-#     happy_path_projection.numba_project_cylinders("XZ")
-#     assert 1 == 1
-
-
-# #@profile
-# def test_large_proj(large_collection, ez_projection):
-#     ez_projection.numba_project_cylinders("XY")
-#     large_collection.project_cylinders("XZ")
-#     assert 1 == 1
-
-
-# #@profile
-# def test_numba_large_proj(large_collection, ez_projection):
-#     ez_projection.numba_project_cylinders("XY")
-#     large_collection.numba_project_cylinders("XZ")
-#     assert 1 == 1
-
-
 # @pytest.mark.report_uss
 # @pytest.mark.report_tracemalloc
 # @pytest.mark.report_duration
