+++ src/canhydro/Cylinder.py
@@ -7,19 +7,10 @@ from dataclasses import dataclass, field
 import numpy as np
 
 from src.canhydro.DataClasses import Projection
-from src.canhydro.geometry import (draw_cyls, get_projection,
-                                   numba_get_projection)
+from src.canhydro.geometry import (draw_cyls, get_projection)
 from src.canhydro.global_vars import qsm_cols
+from src.canhydro.global_vars import logger
 
-# from descartes import PolygonPatch
-# from mpl_toolkits import mplot3d
-
-
-# import time
-# import copy
-# import math
-# import openpyxl
-# import geopandas as geo
 NAME = "Cylinder"
 
 
@@ -32,11 +23,11 @@ def create_cyl(arr: np.array):
 
 
 @dataclass
-class Cylinder:  # (defaultdict):
+class Cylinder:
     cyl_id: int
-    x: np.ndarray[np.float32]  # len 2 array
-    y: np.ndarray[np.float32]  # len 2 array
-    z: np.ndarray[np.float32]  # len 2 array
+    x: np.ndarray[np.float32]
+    y: np.ndarray[np.float32]
+    z: np.ndarray[np.float32]
     radius: np.float32
     length: np.float32
     branch_order: int
@@ -66,9 +57,6 @@ class Cylinder:  # (defaultdict):
     is_stem: bool = False
     is_divide: bool = False
 
-    # #blood for the blood god, software eng for the filter func
-    # class_attrs = self.__get_class_attributes(type(self))
-    # self.__init_instance(class_attrs, kwargs)
     def __repr__(self):
         return f"Cylinder( cyl_id={self.cyl_id}, x={self.x}, y={self.y}, z={self.z}, radius={self.radius}, length={self.length}, branch_order={self.branch_order}, branch_id={self.branch_id}, volume={self.volume}, parent_id={self.parent_id}, reverse_branch_order={self.reverse_branch_order}, segment_id={self.segment_id}"
 
@@ -86,8 +74,7 @@ class Cylinder:  # (defaultdict):
 
         extract = (
             lambda attr: attrs[columns[attr]]
-        )  # pulls a column from the qsm row (attrs) corrosponding to the input attribute
-
+        )  
         self.dx = self.x[1] - self.x[0]
         self.dy = self.y[1] - self.y[0]
         self.dz = self.z[1] - self.z[0]
@@ -114,15 +101,13 @@ class Cylinder:  # (defaultdict):
             else np.arctan(self.dz / np.sqrt(self.dx**2 + self.dy**2))
         )
         self.xy_area = 0
-        # log.info(str(self.__repr__()))
+        logger.debug(str(self.__repr__()))
 
     def get_projection(self, plane="XY"):
         noCirPoints = 360
         tCir = np.linspace(
             0, 2 * np.pi, noCirPoints
         )  # 360 evenly spaced points between 0 - 2pi (radian degrees)
-        a_ortho = np.cos(tCir)  # x coordinates of the points on a circle
-        b_ortho = np.sin(tCir)  # y coordinates of the points on a circle
 
         if plane == "XY":
             magnitude = [self.dx, self.dy, self.dz]
@@ -135,30 +120,6 @@ class Cylinder:  # (defaultdict):
             vector = [np.transpose(self.y), np.transpose(self.z), np.transpose(self.x)]
 
         projection = get_projection(vector, magnitude, self.radius)
-        self.projected_data[plane] = projection
-        if plane == "XY":
-            self.xy_area = self.projected_data["XY"]["area"]
-        return projection["polygon"]
-
-    def numba_get_projection(self, plane="XY"):
-        noCirPoints = 360
-        tCir = np.linspace(
-            0, 2 * np.pi, noCirPoints
-        )  # 360 evenly spaced points between 0 - 2pi (radian degrees)
-        a_ortho = np.cos(tCir)  # x coordinates of the points on a circle
-        b_ortho = np.sin(tCir)  # y coordinates of the points on a circle
-
-        if plane == "XY":
-            magnitude = [self.dx, self.dy, self.dz]
-            vector = [np.transpose(self.x), np.transpose(self.y), np.transpose(self.z)]
-        elif plane == "XZ":
-            magnitude = [self.dx, self.dz, self.dy]
-            vector = [np.transpose(self.x), np.transpose(self.z), np.transpose(self.y)]
-        else:
-            magnitude = [self.dy, self.dz, self.dx]
-            vector = [np.transpose(self.y), np.transpose(self.z), np.transpose(self.x)]
-
-        projection = numba_get_projection(vector, magnitude, self.radius)
         self.projected_data[plane] = projection
         if plane == "XY":
             self.xy_area = self.projected_data["XY"]["area"]
