+++ src/canhydro/utils.py
@@ -4,44 +4,9 @@ import csv
 import os
 import shutil
 import stat
-from typing import Union
 
 import numpy as np
-from numba import njit, prange
-from numba.typed import List
-
 from src.canhydro.global_vars import input_dir, log, output_dir, time_stamp
-
-
-def stack(to_stack:list[np.array], col: bool == True):
-    """
-        A wrapper for njit stack that handles errors and allows for 
-        less strict typing 
-    """
-    list_of_array = List(to_stack)
-    try:
-        njit_stack(list_of_array,col)
-    except ValueError as err:
-        left_shape = list_of_array[0].shape[0]
-        right_shape = list_of_array[1].shape[0]
-        stack_type ='column' if col else 'row'
-        msg = f'{err}: Cannot {stack_type} stack arrays with shapes {left_shape} and {right_shape}'
-        log.error(msg)
-        raise ValueError(msg) from err
-
-@njit()
-def njit_stack(list_of_array:np.array[np.array()], col: bool):
-    """
-    numba doesn't play well with np stacks, so I had to do it myself
-    """
-    num_in = len(list_of_array)
-    left_shape = list_of_array[0].shape[0]
-    shape = (num_in, left_shape)
-    stacked_array = np.empty(shape)
-    for j in prange(len(list_of_array)): 
-        stacked_array[j] = list_of_array[j]
-    return stacked_array if not col else stacked_array.T
-
 
 def on_rm_error(func, path, exc_info):
     # path contains the path of the file that couldn't be removed
@@ -135,7 +100,9 @@ def intermitent_log(prog: int, whole: int, msg: str, freq: int = 0.0001):
 
 
 def lam_filter(objects, a_lambda: function, return_all: bool = False):
-    """Takes in a lambda that filters on cylinder attrs"""
+    """Takes in a lambda that filters on cylinder attrs
+        returns a list of cylinders meeting that requirements 
+    """
     if a_lambda.__code__.co_name != "<lambda>":
         raise ValueError("a lambda required")
     if return_all:
